refactor: split region map selection and transport/feature parts (#3878)
Extract tile selection/hover/glow painting, transport/feature terrain
overlays, and visibility halo/label helpers into dedicated part files
for Phase 3 flame map modularization.
