+++ app/lib/features/game/flame/region_map/region_map_component_shared_visibility_halos.dart
@@ -1,0 +1,66 @@
+part of 'region_map_component.dart';
+
+/// True when `(x, y)` is within Chebyshev distance ≤ 2 of a fleet marker that
+/// applies the naval move-draft reveal halo. SPEC/ui/map-widget.md.
+bool isCellUnderFleetRevealHalo({
+  required int x,
+  required int y,
+  required List<FleetTileMarkerView> fleetTileMarkers,
+}) {
+  for (final m in fleetTileMarkers) {
+    if (!m.applyFleetRevealHalo) {
+      continue;
+    }
+    if (math.max((x - m.x).abs(), (y - m.y).abs()) <= 2) {
+      return true;
+    }
+  }
+  return false;
+}
+
+/// True when `(x, y)` is within Chebyshev distance <= 2 of a civilian marker
+/// that applies the draft assignment reveal halo.
+bool isCellUnderCivilianRevealHalo({
+  required int x,
+  required int y,
+  required List<CivilianTileMarkerView> civilianTileMarkers,
+}) {
+  for (final m in civilianTileMarkers) {
+    if (!m.applyCivilianRevealHalo) {
+      continue;
+    }
+    if (math.max((x - m.x).abs(), (y - m.y).abs()) <= 2) {
+      return true;
+    }
+  }
+  return false;
+}
+
+/// Effective terrain visibility for map painting (fog + reveal halos).
+///
+/// Used by the region map renderer and tests for sea zone label gating (#1756).
+TileVisibility visibilityForTerrainForMapCell({
+  required CtMapVisibilityMode visibilityMode,
+  required CellViewData cell,
+  required List<FleetTileMarkerView> fleetTileMarkers,
+  required List<CivilianTileMarkerView> civilianTileMarkers,
+}) {
+  if (visibilityMode != CtMapVisibilityMode.playerConstrained) {
+    return cell.visibility;
+  }
+  if (isCellUnderFleetRevealHalo(
+    x: cell.x,
+    y: cell.y,
+    fleetTileMarkers: fleetTileMarkers,
+  )) {
+    return TileVisibility.visible;
+  }
+  if (isCellUnderCivilianRevealHalo(
+    x: cell.x,
+    y: cell.y,
+    civilianTileMarkers: civilianTileMarkers,
+  )) {
+    return TileVisibility.visible;
+  }
+  return cell.visibility;
+}
